Preemptive refactor.
Brochure will have methods like 'Action Brochure::action(ActionConstraint)' etc.
Only fetch what is necessary from the GameDB for iterators, etc. Usually a lot
of the information is redundant: all Action constraints will pointer to the same
Action so fetching the Action for each constraint is pointless. Similarly,
usually we won't care about the underlying ActionDefinition or underlying
ResourceType when using the goal planner. And so on.

If the user wants the information, simply use the (NYI) methods.